Fetch attributes lazily.
Previously, we assumed Attributes for a single file was always
available, which was used to power services such as definition provider.
This works no longer since the active file can change.  This commit
reorganizes the codebase so that attributes are lazily fetched by uri.
We currently don't cache attributes, but that might be a good idea
eventually.