Use proto map for references.
It's pretty neat that protobuf supports maps.

I'm not entirely sold if this is a good idea or not. The benefit is that
the storage format maps directly to how the data is used, no duplicate
filnames are stored. On the other hand, I think `Ranges.ranges` is kind
of an awkward interface.