Create metabrowse-server module to serve sources on-demand
The metabrowse-cli module is great because it generates a static site can
be hosted by any static file server. However, if we control the server
then we can serve the SemanticDBs on-demand for any
classpath+sources.jar.

This commit implements a server that uses the same monaco/metabrowse
JavaScript client but serves SemanticDBs and symbol indexes on-demand.
This enables the server to start fast (a few seconds even for larger
classpaths) and does not requires the classpath/sources to be compiled
with SemanticDB beforehand.

To start the server and browse Scalameta sources with the
scalamacros:paradise compiler plugin enabled:
```
metabrowse-server macroparadise -- org.scalameta:scalameta_2.12:4.0.0
```

This commit implements basic selenium tests to check that it works as
expected. I have no experience whatsoever writing selenium tests so it's
quite hacky (Thread.sleep), but I think it's better than having no tests
at all.  For now, I am happy to not run the selenium tests in CI but
keep them around for manual local testing.